delete before add check entity exist

+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/UdfFuncService.java
@@ -303,7 +303,12 @@ public class UdfFuncService extends BaseService{
     @Transactional(rollbackFor = Exception.class)
     public Result delete(int id) {
         Result result = new Result();
-
+        //check exist
+        UdfFunc udfFunc = udfFuncMapper.selectById(id);
+        if (udfFunc == null) {
+            putMsg(result, Status.RESOURCE_NOT_EXIST);
+            return result;
+        }
         udfFuncMapper.deleteById(id);
         udfUserMapper.deleteByUdfFuncId(id);
         putMsg(result, Status.SUCCESS);
